feat(app): hide header buttons after 3 seconds of mouse inactivity
before it would only be hidden if the mouse wasn't on the page.
however this wont work for single monitor setups if you just want it displayed on the single screen.
now, a useEffect is added and when the mouse moves, using an eventListener we call the handleMouseMove fn, which resets the timer of inactivity.